+++ client/src/pages/Home.jsx
@@ -1,0 +1,12 @@
+import { useQuery } from '@apollo/client';
+
+const Home = () => {
+
+  return (
+    <main className='stndrd-page'>
+      <div><h5>Home</h5></div>
+    </main>
+  );
+};
+
+export default Home;