+++ client/src/pages/Session.jsx
@@ -1,0 +1,27 @@
+import { useQuery } from '@apollo/client';
+
+
+
+const Session = () => {
+  return (
+    <main className='stndrd-page'>
+      <div>
+        <h5>Imaging Sessions</h5>
+
+
+      </div>
+    </main>
+  );
+};
+
+export default Session;
+
+
+
+
+
+
+
+
+
+